remove nsec.app method from nostr-login when ios<17 detected
