+++ apps/fanfares/src/app/components/AppController.tsx
@@ -17,6 +17,7 @@ import {
   usePodcastsUnlocked,
 } from "../controllers/state/podcast-slice"
 import { toast } from "react-toastify"
+import { detectIOSFirmware } from "./MobileDetection"
 
 export interface AppControllerProps {
   children: React.ReactNode
@@ -44,10 +45,17 @@ export function AppController(props: AppControllerProps) {
 
   // absolutely bizarre hack needed because nostr-login references 'document' and that breaks server-side rendering
   useEffect(() => {
+    // detect iOS and disable nsec.app login method if < iOS 17
+    const ios = detectIOSFirmware()
+    const nsecAppMethod: string[] = ios && ios < 17 ? [] : ["connect"]
+
+    console.log(nsecAppMethod)
+
     import("nostr-login")
       .then(async ({ init }) => {
         init({
           bunkers: "nsec.app,login.fanfares.io",
+          methods: ["readOnly", "extension", "local", ...nsecAppMethod],
         })
       })
       .catch(error => console.log("Failed to load nostr-login", error))
